requirements.txt
numpy==1.24.3
pandas==2.0.2
matplotlib==3.7.1
scipy==1.10.1
seaborn==0.12.2
psutil==5.9.5
numba==0.57.1
Pillow==9.5.0
openpyxl==3.1.2
xlrd==2.0.1
scikit-learn==1.2.2 